port some tests from fraction-checks.test.ts, remove old fraction checks
