--- packages/step-checker/src/checks/__tests__/fraction-checks.test.ts
@@ -1,766 +1,0 @@
-import {serializer} from "@math-blocks/semantic";
-import {parse} from "@math-blocks/text-parser";
-
-import StepChecker from "../../step-checker";
-import {Context} from "../../types";
-
-import {checkStep, toHaveMessages, toHaveStepsLike} from "../test-util";
-import {mulOne} from "../axiom-checks";
-import {divIsMulByOneOver, mulInverse, divBySame} from "../fraction-checks";
-import {
-    numberCheck,
-    identifierCheck,
-    exactMatch,
-    checkArgs,
-} from "../basic-checks";
-
-expect.addSnapshotSerializer(serializer);
-expect.extend({toHaveMessages, toHaveStepsLike});
-
-describe.skip("Fraction checks", () => {
-    it("a * 1/b -> a / b", () => {
-        const result = checkStep("a * 1/b", "a / b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    it("a * 1/b * c -> a/b * c", () => {
-        const result = checkStep("a * 1/b * c", "a/b * c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    it("a * b * 1/c -> a * b/c", () => {
-        const result = checkStep("a * b * 1/c", "a * b/c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    it("a * b * 1/c -> a/c * b", () => {
-        const result = checkStep("a * b * 1/c", "a/c * b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "commutative property",
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    it("a * b * 1/c -> ab / c", () => {
-        const result = checkStep("a * b * 1/c", "ab / c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying fractions",
-            "multiplication with identity",
-        ]);
-
-        // TODO: handle this differently depending on whether the multiplication
-        // is explicit or implicit, e.g.
-        // a * b * 1/c -> a * b/c vs. (a)(b)(1/c) -> ab / c
-        expect(result).toHaveStepsLike([
-            ["a * b * 1/c", "(a)(b)(1) / c"],
-            ["(a)(b)(1)", "ab"],
-        ]);
-    });
-
-    it("1/b * a -> a / b", () => {
-        const result = checkStep("1/b * a", "a / b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "commutative property",
-            "multiplying by one over something results in a fraction",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["1/b * a", "a * 1/b"],
-            ["a * 1/b", "a / b"],
-        ]);
-    });
-
-    it("a / b -> a * 1/b", () => {
-        const result = checkStep("a / b", "a * 1/b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-        ]);
-    });
-
-    // TODO: simplify this case
-    it("1/a * 1/b -> (1)(1) / ab -> 1 / ab", () => {
-        const result = checkStep("1/a * 1/b", "1 / ab");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying fractions",
-            "multiplication with identity",
-        ]);
-    });
-
-    it("1 -> a/a", () => {
-        const result = checkStep("1", "a/a");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["division by the same value"]);
-    });
-
-    it("a/a -> 1", () => {
-        const result = checkStep("a/a", "1");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["division by the same value"]);
-    });
-
-    describe("divBySame is equivalent to [divIsMulByOneOver, mulInverse]", () => {
-        const checks = [
-            // basic checks
-            numberCheck,
-            identifierCheck,
-            exactMatch,
-
-            // We do this after axiom checks so that we can include commute steps
-            // first and then check if there's an exact match.  checkArgs ignores
-            // ordering of args so if we ran it first we'd never see any commute
-            // steps in the output.
-            checkArgs,
-
-            divIsMulByOneOver,
-            mulInverse,
-        ];
-        const checker = new StepChecker({skipEvalChecker: true}, checks);
-        const context: Context = {
-            checker,
-            steps: [],
-            successfulChecks: new Set(),
-            reversed: false,
-            mistakes: [],
-        };
-
-        it("1 -> a/a", () => {
-            const result = checker.checkStep(parse("1"), parse("a/a"), context);
-
-            expect(result).toBeTruthy();
-            if (!result) {
-                throw new Error("result is undefind");
-            }
-            expect(result).toHaveMessages([
-                "multiplying the inverse",
-                "multiplying by one over something results in a fraction",
-            ]);
-        });
-
-        it("a/a -> 1", () => {
-            const result = checker.checkStep(parse("a/a"), parse("1"), context);
-
-            expect(result).toBeTruthy();
-            if (!result) {
-                throw new Error("result is undefind");
-            }
-            expect(result).toHaveMessages([
-                "fraction is the same as multiplying by one over",
-                "multiplying the inverse",
-            ]);
-        });
-    });
-
-    // This is essentially subsitution property, we know a/a is equal to 1 so
-    // we replace a/a with 1 and then replace 1 with b/b.
-    it("a/a -> b/b", () => {
-        const result = checkStep("a/a", "b/b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "division by the same value",
-            "division by the same value",
-        ]);
-    });
-
-    // This is one of the test cases that blew up when removing 'divBySame'
-    // TODO: make it work with 'mulInverse' instead of 'divBySame'
-    it("b(a/b) -> a", () => {
-        const result = checkStep("b(a/b)", "a");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-            "multiplying the inverse",
-            "multiplication with identity",
-        ]);
-    });
-
-    it("a * b * 1/b -> a", () => {
-        const result = checkStep("a * b * 1/b", "a");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying the inverse",
-            "multiplication with identity",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a * b * 1/b", "a * 1"],
-            ["a * 1", "a"],
-        ]);
-    });
-
-    it("a * b * 1/a -> b", () => {
-        const result = checkStep("a * b * 1/a", "b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying the inverse",
-            "multiplication with identity",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a * b * 1/a", "1 * b"],
-            ["1 * b", "b"],
-        ]);
-    });
-
-    it("a * b * c * 1/a * 1/c -> b", () => {
-        const result = checkStep("a * b * c * 1/a * 1/c", "b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying the inverse",
-            "multiplication with identity",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a * b * c * 1/a * 1/c", "1 * b * 1"],
-            ["1 * b * 1", "b"],
-        ]);
-    });
-
-    it("a * b/c -> ab / b", () => {
-        const result = checkStep("a * b/c", "ab / c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["multiplying fractions"]);
-    });
-
-    it("a/b * c/d -> ac / bd", () => {
-        const result = checkStep("a/b * c/d", "ac / bd");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["multiplying fractions"]);
-    });
-
-    it("ac / bd -> a/b * c/d", () => {
-        const result = checkStep("ac / bd", "a/b * c/d");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["multiplying fractions"]);
-
-        expect(result).toHaveStepsLike([["ac / bd", "a/b * c/d"]]);
-    });
-
-    it("ab/cd * e/f -> abe / cdf", () => {
-        const result = checkStep("ab/cd * e/f", "abe / cdf");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["multiplying fractions"]);
-    });
-
-    it("a/b * 1/d -> a*1 / bd -> a / bd", () => {
-        const result = checkStep("a/b * 1/d", "a / bd");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplying fractions",
-            "multiplication with identity",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a/b * 1/d", "(a)(1) / bd"],
-            ["(a)(1)", "a"],
-        ]);
-    });
-
-    it("30 / 6 -> 2*3*5 / 2*3 -> 2*3/2*3 * 5/1 -> 1 * 5/1 -> 5/1 -> 5", () => {
-        const checker = new StepChecker({skipEvalChecker: true});
-        const result = checker.checkStep(parse("30 / 6"), parse("5"), {
-            checker,
-            steps: [],
-            successfulChecks: new Set(),
-            reversed: false,
-            mistakes: [],
-        });
-
-        if (!result) {
-            throw new Error("failure");
-        }
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "prime factorization",
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-            "division by one",
-        ]);
-    });
-
-    it("24 / 6 -> 2*2*2*3 / 2*3 -> 2*3/2*3 * 2*2/1 -> 1 * 2*2/1 -> 2*2/1 -> 4/1 -> 4", () => {
-        const checker = new StepChecker({evalFractions: false});
-        const result = checker.checkStep(parse("24 / 6"), parse("4"), {
-            checker,
-            steps: [],
-            successfulChecks: new Set(),
-            reversed: false,
-            mistakes: [],
-        });
-
-        if (!result) {
-            throw new Error("failure");
-        }
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "prime factorization",
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-            "division by one",
-            "evaluation of multiplication",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["24 / 6", "(2*2*2*3) / (2*3)"],
-            ["(2*2*2*3) / (2*3)", "(2)(3)/(2)(3) * (2)(2)/1"],
-            ["(2)(3) / (2)(3)", "1"],
-            ["1 * (2)(2) / 1", "(2 * 2) / 1"],
-            ["(2 * 2) / 1", "2 * 2"],
-            ["2 * 2", "4"],
-        ]);
-    });
-
-    it("(2)(2)(2)(3) / (2)(3) -> (2)(2)(2) / (2)", () => {
-        const checker = new StepChecker({evalFractions: false});
-        const result = checker.checkStep(
-            parse("(2)(2)(2)(3) / (2)(3)"),
-            parse("(2)(2)(2) / 2"),
-            {
-                checker,
-                steps: [],
-                successfulChecks: new Set(),
-                reversed: false,
-                mistakes: [],
-            },
-        );
-
-        if (!result) {
-            throw new Error("failure");
-        }
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-        ]);
-    });
-
-    describe("reciprocals", () => {
-        it("a / (b/c) -> a * c/b", () => {
-            const result = checkStep("a / (b/c)", "a * c/b");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "dividing by a fraction is the same as multiplying by the reciprocal",
-            ]);
-        });
-
-        it("a * c/b -> a / (b/c)", () => {
-            const result = checkStep("a * c/b", "a / (b/c)");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "dividing by a fraction is the same as multiplying by the reciprocal",
-            ]);
-
-            expect(result).toHaveStepsLike([["a * c/b", "a / (b/c)"]]);
-        });
-
-        it("1 / (a/b) -> b / a", () => {
-            const result = checkStep("1 / (a/b)", "b / a");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "dividing by a fraction is the same as multiplying by the reciprocal",
-                "multiplication with identity",
-            ]);
-
-            expect(result).toHaveStepsLike([
-                ["1 / (a/b)", "1 * b/a"],
-                ["1 * b/a", "b / a"],
-            ]);
-        });
-
-        it("1 / (1/a) -> a", () => {
-            const result = checkStep("1 / (1/a)", "a");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "dividing by a fraction is the same as multiplying by the reciprocal",
-                "multiplication with identity",
-                "division by one",
-            ]);
-
-            expect(result).toHaveStepsLike([
-                ["1 / (1/a)", "1 * a/1"],
-                ["1 * a/1", "a/1"], // TODO: fix these nodes?
-                ["a/1", "a"],
-            ]);
-        });
-
-        it("a / (1/b) -> a * b/1 -> ab", () => {
-            const result = checkStep("a / (1/b)", "ab");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "dividing by a fraction is the same as multiplying by the reciprocal",
-                "division by one",
-            ]);
-
-            expect(result).toHaveStepsLike([
-                ["a / (1/b)", "a * b/1"],
-                ["b/1", "b"],
-            ]);
-        });
-
-        it("a/b * b/a -> ab/ba -> ab/ab -> 1", () => {
-            const result = checkStep("a/b * b/a", "1");
-
-            expect(result).toBeTruthy();
-            expect(result).toHaveMessages([
-                "multiplying fractions",
-                "commutative property",
-                "division by the same value",
-            ]);
-
-            expect(result).toHaveStepsLike([
-                ["a/b * b/a", "ab/ba"],
-                ["ab", "ba"],
-                ["ba/ba", "1"],
-            ]);
-        });
-    });
-
-    it("24ab / 6a -> 2*2*2*3*a*b / 2*3*a -> 2*3*a/2*3*a * 2*2/1 -> 1 * 2*2/1 -> 2*2/1 -> 4/1 -> 4b", () => {
-        const result = checkStep("24ab / 6a", "4b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "prime factorization",
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-            "division by one",
-            "evaluation of multiplication",
-        ]);
-    });
-
-    // TODO: make this 2a/a -> a/a * 2 instead
-    it("2a/a -> a/a * 2/1 -> 1 * 2/1 -> 2/1 -> 2", () => {
-        const result = checkStep("2a/a", "2");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-            "division by one",
-        ]);
-    });
-
-    it("2a/a -> 2b [incorrect]", () => {
-        expect(() => checkStep("2a/a", "2b")).toThrow();
-    });
-
-    it("2abc/ab -> ab/ab * 2c/1 -> 1 * 2c/1 -> 2c", () => {
-        const result = checkStep("2abc/ab", "2c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-            "division by one",
-        ]);
-    });
-
-    // test that we don't cancel all common factors
-    it("2abc/ab -> a/a * 2bc/b -> 1 * 2bc/b -> 2bc/b", () => {
-        const result = checkStep("2abc/ab", "2bc/b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-        ]);
-    });
-
-    it("2abc/abd -> 2c/d", () => {
-        const result = checkStep("2abc/abd", "2c/d");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-        ]);
-    });
-
-    // TODO: simplify the steps in this test case
-    it("ab/abde -> ab/ab * 1/de -> 1 * 1/de -> 1/de", () => {
-        const result = checkStep("ab/abde", "1/de");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "extract common factors from numerator and denominator",
-            "division by the same value",
-            "multiplication with identity",
-        ]);
-    });
-
-    it("a * b/b -> a * 1 -> a", () => {
-        const result = checkStep("a * b/b", "a");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "division by the same value",
-            "multiplication with identity",
-        ]);
-    });
-
-    it("a -> a * 1 -> a * b/b", () => {
-        const result = checkStep("a", "a * b/b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "multiplication with identity",
-            "division by the same value",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a", "a * 1"],
-            ["1", "b/b"],
-        ]);
-    });
-
-    it("a -> a / 1", () => {
-        const result = checkStep("a", "a / 1");
-
-        expect(result).toBeTruthy();
-
-        // TODO: provide a way to group common steps together.
-        // These three steps in this order is "division by one"
-        expect(result).toHaveMessages(["division by one"]);
-    });
-
-    it("a / 1 -> a", () => {
-        const result = checkStep("a / 1", "a");
-
-        expect(result).toBeTruthy();
-
-        // TODO: provide a way to group common steps together.
-        // These three steps in this order is "division by one"
-        expect(result).toHaveMessages(["division by one"]);
-    });
-
-    describe("divByOne is equivalent to [divIsMulByOneOver, divBySame, mulOne]", () => {
-        const checks = [
-            // basic checks
-            numberCheck,
-            identifierCheck,
-            exactMatch,
-
-            // We do this after axiom checks so that we can include commute steps
-            // first and then check if there's an exact match.  checkArgs ignores
-            // ordering of args so if we ran it first we'd never see any commute
-            // steps in the output.
-            checkArgs,
-
-            divIsMulByOneOver,
-            divBySame,
-            mulOne,
-        ];
-        const checker = new StepChecker({skipEvalChecker: true}, checks);
-        const context: Context = {
-            checker,
-            steps: [],
-            successfulChecks: new Set(),
-            reversed: false,
-            mistakes: [],
-        };
-
-        it("a -> a / 1", () => {
-            const result = checker.checkStep(
-                parse("a"),
-                parse("a / 1"),
-                context,
-            );
-
-            expect(result).toBeTruthy();
-
-            // TODO: provide a way to group common steps together.
-            // These three steps in this order is "division by one"
-            if (!result) {
-                throw new Error("result not defined");
-            }
-            expect(result).toHaveMessages([
-                "multiplication with identity",
-                "division by the same value",
-                "multiplying by one over something results in a fraction",
-            ]);
-        });
-
-        it("a / 1 -> a", () => {
-            const result = checker.checkStep(
-                parse("a / 1"),
-                parse("a"),
-                context,
-            );
-
-            expect(result).toBeTruthy();
-
-            // TODO: provide a way to group common steps together.
-            // These three steps in this order is "division by one"
-            if (!result) {
-                throw new Error("result not defined");
-            }
-            expect(result).toHaveMessages([
-                "fraction is the same as multiplying by one over",
-                "division by the same value",
-                "multiplication with identity",
-            ]);
-        });
-    });
-
-    it("ab -> ab / 1", () => {
-        const result = checkStep("ab", "ab / 1");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["division by one"]);
-
-        expect(result).toHaveStepsLike([["ab", "ab / 1"]]);
-    });
-
-    // TODO: make sure distribution is including substeps
-    // e.g. 1/c * a + 1/c * b
-    it("(a + b) * 1/c -> a/c + b/c", () => {
-        const result = checkStep("(a + b) * 1/c", "a/c  + b/c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "distribution",
-            "multiplying by one over something results in a fraction",
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    // e.g. 1/c * a + 1/c * b
-    it("a/c + b/c -> 1/c * (a + b)", () => {
-        const result = checkStep("a/c + b/c", "1/c * (a + b)");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-            "commutative property",
-            "fraction is the same as multiplying by one over",
-            "commutative property",
-            "factoring",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a/c", "a * 1/c"],
-            ["a * 1/c", "1/c * a"],
-            ["b/c", "b * 1/c"],
-            ["b * 1/c", "1/c * b"],
-            ["1/c * a + 1/c * b", "1/c * (a + b)"],
-        ]);
-    });
-
-    it("a/c -> 1/c * a", () => {
-        const result = checkStep("a/c", "1/c * a");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-            "commutative property",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a/c", "a * 1/c"],
-            ["a * 1/c", "1/c * a"],
-        ]);
-    });
-
-    it("a/1 + b/1 -> a + b", () => {
-        const result = checkStep("a/1 + b/1", "a + b");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages(["division by one", "division by one"]);
-
-        expect(result).toHaveStepsLike([
-            ["a/1", "a"],
-            ["b/1", "b"],
-        ]);
-    });
-
-    it("(a + b) / c -> (a + b) * 1/c -> a * 1/c + b * 1/c -> a/c + b/c", () => {
-        const result = checkStep("(a + b) / c", "a/c + b/c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-            "distribution",
-            "multiplying by one over something results in a fraction",
-            "multiplying by one over something results in a fraction",
-        ]);
-    });
-
-    it("(a + b) / c -> (a + b) * 1/c", () => {
-        const result = checkStep("(a + b) / c", "(a + b) * 1/c");
-
-        expect(result).toBeTruthy();
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-        ]);
-    });
-
-    it("a/c + b/c -> (a + b) / c", () => {
-        const result = checkStep("a/c  + b/c", "(a + b) / c");
-
-        expect(result).toBeTruthy();
-        expect(result.steps).toHaveLength(4);
-
-        expect(result).toHaveMessages([
-            "fraction is the same as multiplying by one over",
-            "fraction is the same as multiplying by one over",
-            "factoring",
-            "multiplying by one over something results in a fraction",
-        ]);
-
-        expect(result).toHaveStepsLike([
-            ["a / c", "a * 1/c"],
-            ["b / c", "b * 1/c"],
-            ["a * 1/c + b * 1/c", "(a + b) * 1/c"],
-            ["(a + b) * 1/c", "(a + b)/c"],
-        ]);
-    });
-});
